requirements.txt
buttplug-py
keyboard
mouse
pywin32==308 ; sys_platform == 'win32'